+++ parsec/cli.py
@@ -29,7 +29,7 @@ def cli():
 cli.add_command(core_cmd, "core")
 cli.add_command(backend_cmd, "backend")
 
-# Add support for PARSEC_CMD_ARGS env var
+# Add support for GUARDATA_CMD_ARGS env var
 
 vanilla_cli_main = cli.main
 
@@ -38,7 +38,7 @@ def patched_cli_main(args=None, **kwargs):
     if args is None:
         args = sys.argv[1:]
 
-    raw_extra_args = os.environ.get("PARSEC_CMD_ARGS", "")
+    raw_extra_args = os.environ.get("GUARDATA_CMD_ARGS", "")
     args += [os.path.expandvars(x) for x in raw_extra_args.split()]
 
     return vanilla_cli_main(args=args, **kwargs)