+++ parsec/core/gui/systray.py
@@ -16,13 +16,13 @@ class Systray(QSystemTrayIcon):
 
         self.menu = QMenu()
         self.show_action = self.menu.addAction(_("ACTION_SYSTRAY_SHOW_WINDOW"))
-        self.close_action = self.menu.addAction(_("ACTION_SYSTRAY_QUIT_PARSEC"))
+        self.close_action = self.menu.addAction(_("ACTION_SYSTRAY_QUIT_guardata"))
 
         self.on_show = self.show_action.triggered
         self.on_close = self.close_action.triggered
 
         self.setContextMenu(self.menu)
-        self.setIcon(QIcon(":/logos/images/icons/parsec.png"))
+        self.setIcon(QIcon(":/logos/images/icons/guardata.png"))
         self.activated.connect(self.on_activated)
         self.show()