+++ tests/core/gui/test_bootstrap_organization.py
@@ -68,7 +68,7 @@ async def test_bootstrap_organization(aqtbot, backend, gui_ready_for_bootstrap, 
                 "",
                 "You organization <b>NewOrg</b> has been created!<br />\n<br />\n"
                 "You will now be automatically logged in.<br />\n<br />\n"
-                "To help you start with PARSEC, you can read the "
+                "To help you start with guardata, you can read the "
                 '<a href="https://docs.parsec.cloud/en/stable/" title="User guide">user guide</a>.',
             )
         ]