+++ HISTORY.rst
@@ -589,7 +589,7 @@ Parsec 1.1.1 (2019-10-21)
 Bugfixes
 ~~~~~~~~
 
-* Fix argument parsing in backend cli commands (``PARSEC_CMD_ARGS`` env var, db
+* Fix argument parsing in backend cli commands (``GUARDATA_CMD_ARGS`` env var, db
   param and S3 entry point default value) (`#749
   <https://github.com/Scille/parsec-cloud/issues/749>`__)